Update the filters' code ... and fix the relationship between BW and Q for the shelving EQs
Note:  For EQs in sforzando, the `hsh` and `lsh` values for `eqN_type` appear ignored. For the `filN_type` values `lsh` and `hsh`, the filters are indeed shelves, but a factor 0.5 appears to be applied on the resonance (i.e. the bandwidth of the shelves is twice that of sfizz). As far as I can tell our implem follows the RBJ's biquad cookbook properly though, so I left it as is.