Centered, full-width background image
After a previous UI change, the image container had an offset to the left side
with the contained image left-aligned.
This change makes the container full-width and centers the image.
In addition, the piano keyboard has been adapted accordingly, leaving more room for the panels.
The background image size has been increased to 800x300.