Add support for opcodes lotimer/hitimer
While not widely supported or well documented, is very useful especially when
triggering a note with on_locc/on_hicc to prevent a bunch of retriggers.

The region will trigger as long as the elapsed time from the note-on of the most recently triggered region in
the current polyphony group is between lotimer and hitimer amount of seconds. lotimer default 0, hitimer default is
max_float.

Example use: to preventing a region from retriggering too quickly, set lotimer=0.1